Fixed lib. Added according description to the README.md.
